Avoid circular import for query block
