+++ sugar/components/server/cdatamatch.py
@@ -30,7 +30,7 @@ import re
 import sugar.utils.objects
 import sugar.utils.structs
 from sugar.components.server.cdatastore import CDataContainer
-from sugar.components.server.query import QueryBlock
+from sugar.components.server.qelement import QueryBlock
 
 
 class UniformMatch:
